Schema attributeList removed
There is no need for an "attributeList" schema. It is better to have generic
attributes of a data structure (matrix, graph, node, ...) in the top level
of the object instead of having an "attrs" field which contains the generic
attributes: that is unnecessary nesting. At least in this design phase,
it is not possible to specify what kind of generic attributes we allow;
it is easier to allow extra fields in objects.