Unit test documentation + test output
Each unit test should have at least one line documentation on the purpose of
the test. Unit tests test/valid/ from definitions-minimal.json to
event-undo.json were cheched for the description attribute and added when
needed.

File test.sh_output.txt contains the output from test.sh. This is for
converting the test.sh to a Node.js version, when test.sh cannot be run
because of missing ajv-cli. Ajv-cli has to be installed globally, and this
requires sudo, which is too complicated.